refactor(capabilities): simplify command and memory surfaces
